+++ public/app/features/alerting/unified/components/rule-editor/query-and-alert-condition/Query.tsx
@@ -3,31 +3,36 @@ import { useFormContext } from 'react-hook-form';
 
 import { Field, InputControl } from '@grafana/ui';
 
-import { RuleFormType, RuleFormValues } from '../../types/rule-form';
+import { RuleFormType, RuleFormValues } from '../../../types/rule-form';
+import { ExpressionEditor } from '../ExpressionEditor';
+import { QueryEditor } from '../QueryEditor';
 
-import { ExpressionEditor } from './ExpressionEditor';
-import { QueryEditor } from './QueryEditor';
-import { RuleEditorSection } from './RuleEditorSection';
-
-export const QueryStep: FC = () => {
+export const Query: FC = () => {
   const {
     control,
     watch,
     formState: { errors },
   } = useFormContext<RuleFormValues>();
+
   const type = watch('type');
   const dataSourceName = watch('dataSourceName');
 
+  const isGrafanaManagedType = type === RuleFormType.grafana;
+  const isCloudAlertRuleType = type === RuleFormType.cloudAlerting;
+  const isRecordingRuleType = type === RuleFormType.cloudRecording;
+
+  const showCloudExpressionEditor = (isRecordingRuleType || isCloudAlertRuleType) && dataSourceName;
+
   return (
-    <RuleEditorSection
-      stepNo={2}
-      title={type === RuleFormType.cloudRecording ? 'Create a query to be recorded' : 'Create a query to be alerted on'}
-    >
-      {(type === RuleFormType.cloudRecording || type === RuleFormType.cloudAlerting) && dataSourceName && (
+    <div>
+      {/* This is the PromQL Editor for Cloud rules and recording rules */}
+      {showCloudExpressionEditor && (
         <Field error={errors.expression?.message} invalid={!!errors.expression?.message}>
           <InputControl
             name="expression"
-            render={({ field: { ref, ...field } }) => <ExpressionEditor {...field} dataSourceName={dataSourceName} />}
+            render={({ field: { ref, ...field } }) => {
+              return <ExpressionEditor {...field} dataSourceName={dataSourceName} />;
+            }}
             control={control}
             rules={{
               required: { value: true, message: 'A valid expression is required' },
@@ -35,7 +40,9 @@ export const QueryStep: FC = () => {
           />
         </Field>
       )}
-      {type === RuleFormType.grafana && (
+
+      {/* This is the editor for Grafana managed rules */}
+      {isGrafanaManagedType && (
         <Field
           invalid={!!errors.queries}
           error={(!!errors.queries && 'Must provide at least one valid query.') || undefined}
@@ -50,6 +57,6 @@ export const QueryStep: FC = () => {
           />
         </Field>
       )}
-    </RuleEditorSection>
+    </div>
   );
 };
